requirements.txt
Django>=4.2.6, <4.3
django-cors-headers>=4.3.0, <4.4
django-filter>=23.3, <24.0
djangorestframework>=3.14.0, <3.15
gunicorn>=21.2.0, <21.3
psycopg2>=2.9.9, <3.0
python-dotenv>=1.0.0, <1.1
requests>=2.31.0, <2.32
whitenoise>=6.6.0, <6.7
dj-database-url>=2.1.0, <2.3
pdf2image>=1.17.0, <1.18.0
google-cloud-documentai>=2.24.2, < 2.25
google-generativeai>=0.4.1, <0.5
django-q>=1.3.9, <1.4